+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
-## 0.0.1
+## 0.0.2
+- Implemented Repository.
+- Implemented Use cases.
+- Implemented Entity.
+- Registered with Get It Dependency Injection.
 
+## 0.0.1
 - Initial release of `fire_storage_impl`.
 - Added support for uploading files to Firebase Storage from mobile and web.
 - Included toast notification support using `app_toast`.